www/react: Move CSS classes related code to separate file
