+++ www/react-base/src/util/LineCssClasses.ts
@@ -1,0 +1,36 @@
+/*
+  This file is part of Buildbot.  Buildbot is free software: you can
+  redistribute it and/or modify it under the terms of the GNU General Public
+  License as published by the Free Software Foundation, version 2.
+
+  This program is distributed in the hope that it will be useful, but WITHOUT
+  ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+  FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+  details.
+
+  You should have received a copy of the GNU General Public License along with
+  this program; if not, write to the Free Software Foundation, Inc., 51
+  Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+  Copyright Buildbot Team Members
+*/
+
+export type LineCssClasses = {
+  firstPos: number;
+  lastPos: number;
+  cssClasses: string;
+}
+
+export function cssClassesMapToCssString(classes: {[key: string]: boolean}): string {
+  let res = '';
+  let isFirst = true;
+  for (const c in classes) {
+    if (isFirst) {
+      isFirst = false;
+    } else {
+      res += ' ';
+    }
+    res += c;
+  }
+  return res;
+}
